requirements.txt
astroid==1.6.3
autopep8==1.3.5
Django==1.11
django-autoslug==1.9.3
django-betterforms==1.2
django-crispy-forms==1.7.2
django-multiupload==0.5.2
easy-thumbnails==2.5
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.3.12
Pillow==5.0.0
pycodestyle==2.4.0
pylint==1.9.1
pylint-django==0.11.1
pylint-plugin-utils==0.2.6
pytz==2018.4
rope==0.10.7
six==1.11.0
wrapt==1.10.11
